feat(core): introduce bootloader emulator

+++ core/embed/bootloader/bootui.h
@@ -58,7 +58,7 @@ uint32_t ui_screen_wipe_confirm(void);
 void ui_screen_wipe(void);
 void ui_screen_wipe_progress(int pos, int len);
 
-void ui_screen_done(int restart_seconds, secbool full_redraw);
+void ui_screen_done(uint8_t restart_seconds, secbool full_redraw);
 
 void ui_screen_fail(void);
